Changed QHash on QMap

+++ Peer/servermanager.cpp
@@ -12,11 +12,11 @@ ServerManager::ServerManager(QTcpSocket *socket, ApplicationInfo& info)
   AbstractStrategy* reg = new RegisterResponseStrategy();
   AbstractStrategy* friend_update = new FriendUpdateStrategy();
 
-  strategies_.insert(static_cast<quint8>(ServerRequest::LOGIN_SUCCEED), login);
-  strategies_.insert(static_cast<quint8>(ServerRequest::LOGIN_FAILED), login);
-  strategies_.insert(static_cast<quint8>(ServerRequest::REGISTER_SUCCEED), reg);
-  strategies_.insert(static_cast<quint8>(ServerRequest::REGISTER_FAILED), reg);
-  strategies_.insert(static_cast<quint8>(ServerRequest::FRIEND_UPDATE_INFO), friend_update);
+  strategies_.insert(ServerRequest::LOGIN_SUCCEED, login);
+  strategies_.insert(ServerRequest::LOGIN_FAILED, login);
+  strategies_.insert(ServerRequest::REGISTER_SUCCEED, reg);
+  strategies_.insert(ServerRequest::REGISTER_FAILED, reg);
+  strategies_.insert(ServerRequest::FRIEND_UPDATE_INFO, friend_update);
 }
 
 ServerManager::~ServerManager() {}
